Fix rebase issue
Fix chan size
